Move to worker thread style
Really, the reason the batch version was so simple was because the
thread was able to sit and monitor the situation. It didn't require
sharing a whole bunch of state between threads. Given the scope of this
library, adding an extra thread to do just that isn't really an issue.